# packages/api-client/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API liveness health check.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto_Output"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/v1/health/postgres": {
      "get": {
        "operationId": "AppController_postgres",
        "parameters": [],
        "responses": {
          "200": {
            "description": "PostgreSQL health check.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostgresHealthResponseDto_Output"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "PostgreSQL is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/v1/health/mongo": {
      "get": {
        "operationId": "AppController_mongo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "MongoDB audit cluster health check.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MongoHealthResponseDto_Output"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "MongoDB audit cluster is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/v1/health/redis": {
      "get": {
        "operationId": "AppController_redis",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Redis health check.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedisHealthResponseDto_Output"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "Redis is unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/v1/health/queues": {
      "get": {
        "operationId": "AppController_queues",
        "parameters": [],
        "responses": {
          "200": {
            "description": "BullMQ queue health check.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueHealthResponseDto_Output"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "503": {
            "description": "BullMQ queues are unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi OTP đăng ký Passenger (Email).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi Email OTP (Resend/console).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto_Output"
                }
              }
            }
          },
          "429": {
            "description": "Vượt giới hạn OTP.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Xác thực OTP → cấp access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "OTP không hợp lệ.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/operator/login": {
      "post": {
        "operationId": "AuthController_operatorLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Operator/Employee `{slug}/{username}`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Sai thông tin đăng nhập.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Tài khoản bị khóa.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/platform/login": {
      "post": {
        "operationId": "AuthController_platformLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Platform `platform/{username}`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Sai thông tin đăng nhập.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Tài khoản bị khóa.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/oauth/{provider}": {
      "post": {
        "operationId": "AuthController_oauth",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "description": "OAuth provider (v1: google).",
            "schema": {
              "enum": [
                "google"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthInitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Khởi tạo OAuth (v1: Google).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthRedirectResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/oauth/session": {
      "post": {
        "operationId": "AuthController_oauthSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Đổi Better Auth session (sau OAuth callback) → access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Marketplace Ve Xe Nhanh API",
    "description": "Managed marketplace API for passenger, operator, and platform workflows.",
    "version": "0.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ]
      },
      "HealthResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          },
          "service": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "status",
          "service",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "PostgresHealthResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          },
          "service": {
            "type": "string",
            "const": "postgres"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "status",
          "service",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "MongoHealthResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          },
          "service": {
            "type": "string",
            "const": "mongo"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "status",
          "service",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "RedisHealthResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          },
          "service": {
            "type": "string",
            "const": "redis"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "status",
          "service",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "QueueHealthResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          },
          "queues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "counts": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                }
              },
              "required": [
                "name",
                "counts"
              ],
              "additionalProperties": false
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "status",
          "queues",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "MessageResponseDto_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "otp": {
            "type": "string",
            "minLength": 4,
            "maxLength": 10
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "AuthTokenResponseDto_Output": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "const": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "scope": {
            "type": "string",
            "enum": [
              "passenger",
              "operator",
              "platform"
            ]
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn",
          "scope",
          "role"
        ],
        "additionalProperties": false
      },
      "CredentialLoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 3,
            "maxLength": 160
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "OAuthInitDto": {
        "type": "object",
        "properties": {
          "callbackURL": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "OAuthRedirectResponseDto_Output": {
        "type": "object",
        "properties": {
          "redirectUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "redirectUrl"
        ],
        "additionalProperties": false
      }
    }
  }
}